Rip out HtmlFormatter
