+++ tracing-test-utils/src/main/java/com/palantir/tracing/TestTracingExtension.java
@@ -16,17 +16,12 @@
 
 package com.palantir.tracing;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
-import com.palantir.conjure.java.serialization.ObjectMappers;
-import com.palantir.tracing.api.DeserializeSpan;
+import com.palantir.tracing.api.Serialization;
 import com.palantir.tracing.api.Span;
-import java.io.IOException;
-import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 import java.util.function.Function;
@@ -40,7 +35,6 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 final class TestTracingExtension implements BeforeEachCallback, AfterEachCallback {
 
     private final TestTracingSubscriber subscriber = new TestTracingSubscriber();
-    private static final ObjectMapper mapper = ObjectMappers.newServerObjectMapper();
 
     @Override
     public void beforeEach(ExtensionContext context) {
@@ -58,12 +52,12 @@ final class TestTracingExtension implements BeforeEachCallback, AfterEachCallbac
 
         // match recorded traces against expected file (or create)
         if (!Files.exists(file) || Boolean.valueOf(System.getProperty("recreate", "false"))) {
-            serialize(file, subscriber.getAllSpans());
+            Serialization.serialize(file, subscriber.getAllSpans());
             return;
         }
 
         // TODO filter for just one traceId (??) to figure out concurrency
-        SpanAnalyzer.Result expected = SpanAnalyzer.analyze(deserialize(file));
+        SpanAnalyzer.Result expected = SpanAnalyzer.analyze(Serialization.deserialize(file));
         SpanAnalyzer.Result actual = SpanAnalyzer.analyze(subscriber.getAllSpans());
 
         List<String> problemSpanIds = compareSpansRecursively(expected, actual, expected.root(), actual.root())
@@ -109,32 +103,5 @@ final class TestTracingExtension implements BeforeEachCallback, AfterEachCallbac
 
     private static String testName(ExtensionContext context) {
         return context.getRequiredTestClass().getSimpleName() + "/" + context.getRequiredTestMethod().getName();
-    }
-
-    private static void serialize(Path file, Collection<Span> allSpans) throws IOException {
-        Files.createDirectories(file.getParent());
-        try (OutputStream outputStream = Files.newOutputStream(file)) {
-            allSpans.forEach(span -> {
-                try {
-                    byte[] bytes = mapper.writeValueAsBytes(span);
-                    outputStream.write(bytes);
-                    outputStream.write('\n');
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            });
-        }
-    }
-
-    private static List<Span> deserialize(Path file) throws IOException {
-        try (Stream<String> lines = Files.lines(file)) {
-            return lines.map(line -> {
-                try {
-                    return DeserializeSpan.deserialize(mapper, line);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            }).collect(ImmutableList.toImmutableList());
-        }
     }
 }
